added crud for categories, tags and pictures; payment system; pagination

+++ controllers/userController.js
@@ -13,20 +13,9 @@ const factory = require('./handlerFactory');
 const AppError = require('./../utils/appError');
 const catchAsync = require('../utils/catchAsync');
 
-// const multerStorage = multer.diskStorage({
-//   destination: (req, file, cb) => {
-//     cb(null, 'public/imgs/users/');
-//   },
-//   filename: (req, file, cb) => {
-//     const ext = file.mimetype.split('/')[1];
-//     cb(null, `user-${req.user.id}.${ext}`);
-//   },
-// });
-
 const multerStorage = multer.memoryStorage();
 
 const multerFilter = (req, file, cb) => {
-  console.log(file);
   if (file.mimetype.startsWith('image')) cb(null, true);
   else cb(new AppError('Завантажити можна лише фото', 400), false);
 };
@@ -48,19 +37,19 @@ exports.checkPasswordUpdating = (req, res, next) => {
   next();
 };
 
-exports.addAvatar = (req, res, next) => {
+exports.addAvatar = catchAsync(async (req, res, next) => {
   if (!req.file) next();
   else {
     req.file.filename = `${uuidv4()}.jpeg`;
     req.body.avatar = req.file.filename;
-    sharp(req.file.buffer)
+    await sharp(req.file.buffer)
       .resize(500, 500)
       .toFormat('jpeg')
       .jpeg({ quality: 90 })
       .toFile(`public/imgs/users/${req.file.filename}`);
     next();
   }
-};
+});
 
 exports.deleteAvatar = catchAsync(async (req, res, next) => {
   if (req.user.avatar && req.user.avatar != 'default.png')
@@ -75,7 +64,17 @@ exports.deleteAvatar = catchAsync(async (req, res, next) => {
 exports.deleteUserConnects = catchAsync(async (req, res, next) => {
   await Liked.deleteMany({ user_id: req.user.id });
   await Bought.deleteMany({ user_id: req.user.id });
-  await Picture.deleteMany({ artist_id: req.user.id });
+  console.log(req.user);
+
+  if (req.user.role === 'artist') {
+    const pictures = await Picture.find({ artist_id: req.user.id });
+    for (let i = 0; i < pictures.length; i++) {
+      await unlinkAsync(`public/imgs/pictures/demo/${pictures[i].demo}`);
+      await unlinkAsync(`public/imgs/pictures/full/${pictures[i].image}`);
+    }
+    await Picture.deleteMany({ artist_id: req.user.id });
+  }
+
   next();
 });
